pyproject.toml
[tool.poetry]
name = "aitoolkit"
version = "0.1.0"
description = ""
authors = ["aqiu <819110812@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<4.0,>=3.8.1"
langchain = "^0.1.16"
openai = "^1.23.3"
langchainhub = "^0.1.15"
langchain-community = "^0.0.34"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-restful = "^0.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
